refactor(transpiler): unify exception handling for CRD transpilation errors
