+++ src/KubeOps.Transpiler/Exceptions/CircularTypeReferenceException.cs
@@ -2,14 +2,14 @@
 // The .NET Foundation licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information.
 
-namespace KubeOps.Transpiler;
+namespace KubeOps.Transpiler.Exceptions;
 
 /// <summary>
 /// Raised when the CRD transpiler detects a circular type reference that cannot be represented as a
 /// finite OpenAPI schema. Derives from <see cref="InvalidOperationException"/> to preserve backwards
 /// compatibility for existing catch clauses.
 /// </summary>
-public sealed class CircularTypeReferenceException : InvalidOperationException
+internal sealed class CircularTypeReferenceException : InvalidOperationException
 {
     /// <summary>
     /// Initializes a new instance of the <see cref="CircularTypeReferenceException"/> class.
